idalloc: assert against counter regressions
Touches #48485.

Also switches the counter type to int64. The choice of uint32 was
historical and not well-motivated.

Release note: None

+++ pkg/kv/kvserver/idalloc/helpers_test.go
@@ -11,6 +11,6 @@
 package idalloc
 
 // IDs returns the channel that the allocator uses to buffer available ids.
-func (ia *Allocator) IDs() <-chan uint32 {
+func (ia *Allocator) IDs() <-chan int64 {
 	return ia.ids
 }
